Issue #24: Created an in-page audio player. Added basic API methods.

+++ tales_django/entities/Tracks/models/Track.py
@@ -74,7 +74,7 @@ class Track(Model):
 
     for_members = models.BooleanField(_('for members only'), default=False)   # , verbose_name=_('for members only'))
 
-    listened_count = models.BigIntegerField(blank=True, default=0, help_text=_('Listened count'))
+    played_count = models.BigIntegerField(blank=True, default=0, help_text=_('Played count'))
 
     # Properties derived from the audio track file
     audio_duration = models.BigIntegerField(null=True, help_text=_('Duration (seconds)'))
@@ -109,9 +109,19 @@ class Track(Model):
         # Try to remove the old files...
         try:
             track = Track.objects.get(id=self.id)
-            if track.audio_file and self.audio_file and track.audio_file != self.audio_file:
+            if (
+                track.audio_file
+                and self.audio_file
+                and track.audio_file != self.audio_file
+                and not str(track.audio_file).startsWith('samples/')
+            ):
                 track.audio_file.delete(save=False)
-            if track.preview_picture and self.preview_picture and track.preview_picture != self.preview_picture:
+            if (
+                track.preview_picture
+                and self.preview_picture
+                and track.preview_picture != self.preview_picture
+                and not str(track.preview_picture).startsWith('samples/')
+            ):
                 track.preview_picture.delete(save=False)
         except Track.DoesNotExist:
             # Do nothing when a new object is creating
